+++ Cranky/AnalyzerResult.cs
@@ -7,4 +7,6 @@ public record AnalyzerResult(int Total, int Undocumented)
     public double DocumentedPercentage => (double)Documented / Total;
 
     public int DocumentedPercentageDisplay => (int)(DocumentedPercentage * 100.0);
+
+    public Dictionary<string, int> PerFilePercentage = new();
 }